social_media_api/requirements.txt
Django
djangorestframework
gunicorn
psycopg2-binary
dj-database-url